python-backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Client for external APIs
httpx==0.25.2

# AI Integration
openai==1.3.7

# Data Validation and Serialization
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0

# Additional Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
